ビルドエラー対応１

+++ DB/DBB/standard/business/src/main/java/jp/co/ndensan/reams/db/dbb/business/core/basic/KibetsuChoshuYuyo.java
@@ -14,7 +14,6 @@ import jp.co.ndensan.reams.ur.urz.definition.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.ParentModelBase;
 import jp.co.ndensan.reams.uz.uza.util.db.EntityDataState;
 
@@ -40,7 +39,7 @@ public class KibetsuChoshuYuyo extends ParentModelBase<KibetsuChoshuYuyoIdentifi
     public KibetsuChoshuYuyo(FlexibleYear 調定年度,
             FlexibleYear 賦課年度,
             TsuchishoNo 通知書番号,
-            Decimal 履歴番号,
+            int 履歴番号,
             RString 徴収方法,
             int 期) {
         requireNonNull(調定年度, UrSystemErrorMessages.値がnull.getReplacedMessage("調定年度"));
@@ -130,7 +129,7 @@ public class KibetsuChoshuYuyo extends ParentModelBase<KibetsuChoshuYuyoIdentifi
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public int get履歴番号() {
         return entity.getRirekiNo();
     }
