+++ DB/DBB/standard/business/src/test/java/jp/co/ndensan/reams/db/dbb/business/core/basic/ChoshuYuyoBuilderTest.java
@@ -11,7 +11,6 @@ import jp.co.ndensan.reams.db.dbz.definition.core.valueobject.code.kyotsu.Hokenr
 import jp.co.ndensan.reams.db.dbz.definition.core.valueobject.code.kyotsu.HokenryoChoshuYuyoTorikeshiShurui;
 import jp.co.ndensan.reams.db.dbz.testhelper.DbbTestBase;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import org.junit.Before;
@@ -32,7 +31,7 @@ public class ChoshuYuyoBuilderTest extends DbbTestBase {
     private static FlexibleYear 調停年度;
     private static FlexibleYear 賦課年度;
     private static TsuchishoNo 通知書番号;
-    private static Decimal 履歴番号;
+    private static int 履歴番号;
 
     @BeforeClass
     public static void setUpClass() {
