+++ DB/DBB/standard/persistence/src/test/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2006ChoshuYuyoDacTest.java
@@ -15,7 +15,6 @@ import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.TsuchishoNo
 import jp.co.ndensan.reams.db.dbz.testhelper.DbbTestDacBase;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.EntityDataState;
 import jp.co.ndensan.reams.uz.uza.util.di.InstanceProvider;
 import static org.hamcrest.CoreMatchers.is;
@@ -225,7 +224,7 @@ public class DbT2006ChoshuYuyoDacTest extends DbbTestDacBase {
                 FlexibleYear 調定年度,
                 FlexibleYear 賦課年度,
                 TsuchishoNo 通知書番号,
-                Decimal 履歴番号) {
+                int 履歴番号) {
             DbT2006ChoshuYuyoEntity entity = DbT2006ChoshuYuyoEntityGenerator.createDbT2006ChoshuYuyoEntity();
             entity.setChoteiNendo(調定年度);
             entity.setFukaNendo(賦課年度);
